tests/integration: Move `config`command test to a dedicated file
`podman-compose config` command lacked a dedicated integration-test
file.
